aeroassist/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.32.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.9.1
sqlalchemy==2.0.28
openai>=1.93.0
psycopg2-binary==2.9.9